Revert local hack to sources
The sources should be pristine clear and match upstream code bit to bit.
Modifications to the upstream code has to be shipped in the form of
patches under debian/patches, not as local hack. Else the Debian
packaging will complain about local difference.

Remove the custom patches and update the debian/patches/reqfix instead.

Revert
* "fix setup" 98cd21f0b7fda589e72622cb46a9554c9ac3fac8
* "add final newline" 72680b58ffa2dcda240f35b8e87ee3c03206a755
* "Patch requirements in debian branch" b86c1a8d0d0aebbc90ce6c3c3fa09df4e61958ed

Change-Id: Iec9cb2ec8281ba2f0fb137eeb4413a59a17fe1a3